pyproject.toml
[project]
name = "workflow-approvals"
version = "0.1.0"
description = ""
authors = [
    {name = "Jesse Young",email = "jesse.young@databricks.com"}
]
requires-python = "3.12.*"

dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "databricks-sql-connector (>=4.0.3,<5.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
databricks-connect = "==16.3.*"
databricks-sdk = {extras = ["notebook"], version = "^0.51.0"}
ipython = "^9.2.0"
ipykernel = "^6.29.5"

